Rename HighwayContext to ClContext.
Highway works with any `Context`. The concrete selection of key, ID and
block types is specific to the CasperLabs node, not to Highway.
