+++ node/src/components/consensus/protocols/highway.rs
@@ -4,7 +4,6 @@ use std::{
     any::Any,
     collections::{BTreeMap, HashMap},
     fmt::Debug,
-    rc::Rc,
 };
 
 use datasize::DataSize;
@@ -16,7 +15,6 @@ use self::round_success_meter::RoundSuccessMeter;
 
 use crate::{
     components::consensus::{
-        candidate_block::CandidateBlock,
         consensus_protocol::{BlockContext, ConsensusProtocol, ConsensusProtocolResult},
         highway_core::{
             active_validator::Effect as AvEffect,
@@ -27,11 +25,7 @@ use crate::{
             validators::Validators,
             Weight,
         },
-        traits::{Context, NodeIdT, ValidatorSecret},
-    },
-    crypto::{
-        asymmetric_key::{self, PublicKey, SecretKey, Signature},
-        hash::{self, Digest},
+        traits::{Context, NodeIdT},
     },
     types::{CryptoRngCore, Timestamp},
 };
@@ -482,52 +476,5 @@ where
 
     fn as_any(&self) -> &dyn Any {
         self
-    }
-}
-
-pub(crate) struct HighwaySecret {
-    secret_key: Rc<SecretKey>,
-    public_key: PublicKey,
-}
-
-impl HighwaySecret {
-    pub(crate) fn new(secret_key: Rc<SecretKey>, public_key: PublicKey) -> Self {
-        Self {
-            secret_key,
-            public_key,
-        }
-    }
-}
-
-impl ValidatorSecret for HighwaySecret {
-    type Hash = Digest;
-    type Signature = Signature;
-
-    fn sign(&self, hash: &Digest, rng: &mut dyn CryptoRngCore) -> Signature {
-        asymmetric_key::sign(hash, self.secret_key.as_ref(), &self.public_key, rng)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub(crate) struct HighwayContext;
-
-impl Context for HighwayContext {
-    type ConsensusValue = CandidateBlock;
-    type ValidatorId = PublicKey;
-    type ValidatorSecret = HighwaySecret;
-    type Signature = Signature;
-    type Hash = Digest;
-    type InstanceId = Digest;
-
-    fn hash(data: &[u8]) -> Digest {
-        hash::hash(data)
-    }
-
-    fn verify_signature(hash: &Digest, public_key: &PublicKey, signature: &Signature) -> bool {
-        if let Err(error) = asymmetric_key::verify(hash, signature, public_key) {
-            info!(%error, %signature, %public_key, %hash, "failed to validate signature");
-            return false;
-        }
-        true
     }
 }
